qwertigraph/pyproject.toml
[project]
name = "qwertigraph"
version = "0.1.0"
description = "Python text expansion using Gregg-inspired briefs"
authors = [
    {name = "Kevin Knox",email = "kevinknox@embarqmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pynput (>=1.8.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
